+++ core-ui/lib-core-ui/src/main/java/ru/surfstudio/android/core/ui/event/result/BaseActivityResultDelegate.java
@@ -1,5 +1,5 @@
 /*
-  Copyright (c) 2018-present, SurfStudio LLC, Maxim Tuev.
+  Copyright (c) 2018-present, SurfStudio LLC, Maxim Tuev, Maxim Smirnov.
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -18,89 +18,162 @@ package ru.surfstudio.android.core.ui.event.result;
 
 import android.app.Activity;
 import android.content.Intent;
+import android.util.SparseArray;
+
+import androidx.annotation.NonNull;
+import androidx.core.util.Pair;
 
 import java.io.Serializable;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Map;
 import java.util.Set;
 
 import io.reactivex.Observable;
 import io.reactivex.subjects.PublishSubject;
 import io.reactivex.subjects.Subject;
 import ru.surfstudio.android.core.ui.navigation.ScreenResult;
+import ru.surfstudio.android.core.ui.navigation.ScreenResultNoData;
 import ru.surfstudio.android.logger.Logger;
 
 /**
  * базовый класс делегата, позволяющий регистрировать обработчики на событие onActivityResult
  */
 public class BaseActivityResultDelegate implements ActivityResultDelegate {
-    private Map<Integer, ActivityResultRegistration> activityResultSubjects = new HashMap<>();
 
+    private SparseArray<ActivityResultRegistration> activityResultSubjects = new SparseArray<>();
+    private SparseArray<ActivityResultNoDataRegistration> activityResultNoDataSubjects = new SparseArray<>();
+
+    @SuppressWarnings("unchecked")
     @Override
     public boolean onActivityResult(int requestCode, int resultCode, Intent data) {
-        for (Map.Entry<Integer, ActivityResultRegistration> routeEntry : activityResultSubjects.entrySet()) {
-            if (routeEntry.getKey() == requestCode) {
-                Subject resultSubject = routeEntry.getValue().getSubject();
-                resultSubject.onNext(new ScreenResult<>(
-                        resultCode == Activity.RESULT_OK,
-                        data != null ? routeEntry.getValue().getRoute().parseResultIntent(data) : null));
-                return true;
+        boolean result = false;
+
+        int code;
+        for (int i = 0; i < activityResultSubjects.size(); i++) {
+            code = activityResultSubjects.keyAt(i);
+            Pair<Integer, ActivityResultRegistration> routePair = new Pair<>(code, activityResultSubjects.get(code));
+            if (routePair.first != null
+                    && routePair.first == requestCode
+                    && routePair.second != null) {
+                Subject resultSubject = routePair.second.getSubject();
+                resultSubject.onNext(
+                        new ScreenResult(
+                                resultCode == Activity.RESULT_OK,
+                                data != null ? ((SupportOnActivityResultRoute) routePair.second.getRoute())
+                                        .parseResultIntent(data) : null
+                        )
+                );
+                result = true;
             }
         }
-        return false;
+
+        for (int i = 0; i < activityResultNoDataSubjects.size(); i++) {
+            code = activityResultNoDataSubjects.keyAt(i);
+            Pair<Integer, ActivityResultNoDataRegistration> routePair = new Pair<>(code, activityResultNoDataSubjects.get(code));
+            if (routePair.first != null
+                    && routePair.first == requestCode
+                    && routePair.second != null) {
+                Subject resultSubject = routePair.second.getSubject();
+                resultSubject.onNext(
+                        new ScreenResultNoData(resultCode == Activity.RESULT_OK)
+                );
+                result = true;
+            }
+        }
+
+        return result;
     }
 
-    protected <T extends Serializable> Observable<ScreenResult<T>> observeOnActivityResult(
-            SupportOnActivityResultRoute<T> route) {
+    protected boolean isObserved(SupportCodeActivityRoute route) {
+        final int code = route.getRequestCode();
+        return activityResultSubjects.indexOfKey(code) != -1
+                || activityResultNoDataSubjects.indexOfKey(code) != -1;
+    }
+
+    @NonNull
+    protected <T extends Serializable> Observable<ScreenResult<T>> observeOnActivityResult(SupportOnActivityResultRoute<T> route) {
         tryRemoveDuplicateResultSubjects(route);
         PublishSubject<ScreenResult<T>> resultSubject = PublishSubject.create();
-        activityResultSubjects.put(route.getRequestCode(), new ActivityResultRegistration(route, resultSubject));
+        activityResultSubjects.put(route.getRequestCode(), new ActivityResultRegistration<>(route, resultSubject));
         return resultSubject;
     }
 
-    protected boolean isObserved(SupportOnActivityResultRoute route) {
-        for (Integer registeredRequestCode : activityResultSubjects.keySet()) {
-            if (registeredRequestCode == route.getRequestCode()) {
-                return true;
-            }
-        }
-        return false;
+    @NonNull
+    protected Observable<ScreenResultNoData> observeOnActivityResultNoData(SupportOnActivityResultNoDataRoute route) {
+        tryRemoveDuplicateResultSubjects(route);
+        PublishSubject<ScreenResultNoData> resultSubject = PublishSubject.create();
+        activityResultNoDataSubjects.put(route.getRequestCode(), new ActivityResultNoDataRegistration(route, resultSubject));
+        return resultSubject;
     }
 
-    private void tryRemoveDuplicateResultSubjects(SupportOnActivityResultRoute route) {
+    private void tryRemoveDuplicateResultSubjects(SupportCodeActivityRoute route) {
         Set<Integer> toDeleteKeys = new HashSet<>();
-        for (Map.Entry<Integer, ActivityResultRegistration> registeredRouteEntry : activityResultSubjects.entrySet()) {
-            if (registeredRouteEntry.getKey() == route.getRequestCode()) {
-                registeredRouteEntry.getValue().getSubject().onComplete();
-                toDeleteKeys.add(registeredRouteEntry.getKey());
+        Set<Pair<Integer, BaseActivityResultRegistration>> registrations = new HashSet<>();
+
+        int code;
+        for (int i = 0; i < activityResultSubjects.size(); i++) {
+            code = activityResultSubjects.keyAt(i);
+            registrations.add(new Pair<>(code, activityResultSubjects.get(code)));
+        }
+        for (int i = 0; i < activityResultNoDataSubjects.size(); i++) {
+            code = activityResultNoDataSubjects.keyAt(i);
+            registrations.add(new Pair<>(code, activityResultNoDataSubjects.get(code)));
+        }
+        for (Pair<Integer, BaseActivityResultRegistration> registeredRoutePair : registrations) {
+            if (registeredRoutePair.first != null
+                    && registeredRoutePair.first == route.getRequestCode()
+                    && registeredRoutePair.second != null) {
+                registeredRoutePair.second.getSubject().onComplete();
+                toDeleteKeys.add(registeredRoutePair.first);
                 Logger.v(this.getClass().getName(), "duplicate registered SupportOnActivityResultRoute: "
-                        + registeredRouteEntry.getValue().getRoute() + " old route unregistered");
+                        + registeredRoutePair.second.getRoute() + " old route unregistered");
             }
         }
         //удаляем ключи здесь, чтобы не было ConcurrentModificationException
         for (Integer key : toDeleteKeys) {
             activityResultSubjects.remove(key);
+            activityResultNoDataSubjects.remove(key);
         }
     }
 
-    private class ActivityResultRegistration<T extends Serializable> {
-        private SupportOnActivityResultRoute<T> route;
-        private PublishSubject<ScreenResult<T>> subject;
+    private abstract class BaseActivityResultRegistration<S, R extends SupportCodeActivityRoute> {
 
-        public ActivityResultRegistration(SupportOnActivityResultRoute<T> route, PublishSubject<ScreenResult<T>> subject) {
+        @NonNull
+        protected R route;
+
+        @NonNull
+        protected PublishSubject<S> subject;
+
+        protected BaseActivityResultRegistration(@NonNull R route, @NonNull
+                PublishSubject<S> subject) {
             this.route = route;
             this.subject = subject;
         }
 
-        public SupportOnActivityResultRoute<T> getRoute() {
+        @NonNull
+        public R getRoute() {
             return route;
         }
 
-        public PublishSubject<ScreenResult<T>> getSubject() {
+        @NonNull
+        public PublishSubject<S> getSubject() {
             return subject;
         }
     }
 
+    private class ActivityResultRegistration<T extends Serializable>
+            extends BaseActivityResultRegistration<ScreenResult<T>, SupportOnActivityResultRoute<T>> {
 
+        public ActivityResultRegistration(SupportOnActivityResultRoute<T> route, PublishSubject<ScreenResult<T>> subject) {
+            super(route, subject);
+        }
+    }
+
+
+    private class ActivityResultNoDataRegistration
+            extends BaseActivityResultRegistration<ScreenResultNoData, SupportOnActivityResultNoDataRoute> {
+
+        public ActivityResultNoDataRegistration(SupportOnActivityResultNoDataRoute route, PublishSubject<ScreenResultNoData> subject) {
+            super(route, subject);
+        }
+    }
 }