+++ core-ui/lib-core-ui/src/main/java/ru/surfstudio/android/core/ui/navigation/ScreenResultNoData.java
@@ -1,5 +1,5 @@
 /*
-  Copyright (c) 2018-present, SurfStudio LLC, Maxim Tuev.
+  Copyright (c) 2018-present, SurfStudio LLC, Maxim Smirnov.
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -13,23 +13,19 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package ru.surfstudio.android.core.ui.navigation.activity.route;
-
-import android.content.Intent;
-
-import java.io.Serializable;
+package ru.surfstudio.android.core.ui.navigation;
 
 /**
- * см {@link ActivityWithResultRoute}
- *
- * @param <T> тип результата
+ * Результат выполнения экрана без данных
  */
-public abstract class ActivityWithParamsAndResultRoute<T extends Serializable> extends ActivityWithResultRoute<T> {
+public class ScreenResultNoData {
+    private boolean success;
 
-    public ActivityWithParamsAndResultRoute() {
+    public ScreenResultNoData(boolean success) {
+        this.success = success;
     }
 
-    public ActivityWithParamsAndResultRoute(Intent intent) {
+    public boolean isSuccess() {
+        return success;
     }
-
 }