+++ core-ui/sample/src/main/java/ru/surfstudio/android/core/ui/sample/ui/screen/main/MainActivityView.kt
@@ -40,9 +40,15 @@ class MainActivityView : BaseRenderableActivityView<MainScreenModel>() {
         }
     }
 
-    override fun renderInternal(sm: MainScreenModel) {}
+    override fun renderInternal(sm: MainScreenModel) {
+        presenter.startResultActivity()
+    }
 
     override fun getPresenters(): Array<CorePresenter<*>> = arrayOf(presenter)
 
     override fun createConfigurator(): DefaultActivityScreenConfigurator = MainScreenConfigurator(intent)
+
+    fun showToast(message: String) {
+        messageController.show(message)
+    }
 }