+++ core-ui/sample/src/main/java/ru/surfstudio/android/core/ui/sample/ui/screen/result/ResultScreenModel.kt
@@ -1,0 +1,6 @@
+package ru.surfstudio.android.core.ui.sample.ui.screen.result
+
+import ru.surfstudio.android.core.mvp.model.ScreenModel
+
+class ResultScreenModel : ScreenModel() {
+}